+++ src/unpack.rs
@@ -29,7 +29,7 @@ struct Unpacker {
     hashes_file: SlabFile,
     stream_file: SlabFile,
 
-    slabs: BTreeMap<u64, Arc<SlabInfo>>,
+    slabs: BTreeMap<u32, Arc<SlabInfo>>,
 }
 
 impl Unpacker {
@@ -73,7 +73,7 @@ impl Unpacker {
         Ok((input, r))
     }
 
-    fn read_info(&mut self, slab: u64) -> Result<Arc<SlabInfo>> {
+    fn read_info(&mut self, slab: u32) -> Result<Arc<SlabInfo>> {
         // Read the hashes slab
         let hashes = self.hashes_file.read(slab)?;
 
@@ -87,7 +87,7 @@ impl Unpacker {
         Ok(Arc::new(SlabInfo { offsets, data }))
     }
 
-    fn get_info(&mut self, slab: u64) -> Result<Arc<SlabInfo>> {
+    fn get_info(&mut self, slab: u32) -> Result<Arc<SlabInfo>> {
         if let Some(info) = self.slabs.get(&slab) {
             Ok(info.clone())
         } else {
@@ -132,7 +132,7 @@ impl Unpacker {
         let mut unpacker = stream::MappingUnpacker::default();
 
         for s in 0..nr_slabs {
-            let stream_data = self.stream_file.read(s as u64)?;
+            let stream_data = self.stream_file.read(s as u32)?;
             let entries = unpacker.unpack(&stream_data[..])?;
             let nr_entries = entries.len();
